chore(infrastructure): move users-collection to mongodb
