+++ src/kairos/authentication/infrastructure/repository/find-one-by-email.test.ts
@@ -1,6 +1,7 @@
 import { invalidCredentialsError } from '@/kairos/authentication/domain/errors';
 import { createFindOneByEmail } from '@/kairos/authentication/infrastructure/repository/find-one-by-email';
-import { usersCollection } from '@/kairos/shared/persistence/mongodb/mongodb-composition';
+
+import { usersCollection } from '@/kairos/shared/infrastructure/mongodb/collection/users-collection';
 import { integrationTest } from '@tests/__vitest__/integration-test';
 import { describe, expect, it } from 'vitest';
