+++ src/kairos/setup/infrastructure/repository/create-super-admin.test.ts
@@ -1,5 +1,6 @@
 import { createCreateSuperAdmin } from '@/kairos/setup/infrastructure/repository/create-super-admin';
-import { usersCollection } from '@/kairos/shared/persistence/mongodb/mongodb-composition';
+
+import { usersCollection } from '@/kairos/shared/infrastructure/mongodb/collection/users-collection';
 import { integrationTest } from '@tests/__vitest__/integration-test';
 import { describe, expect, test } from 'vitest';
